QuantTradePro/requirements.txt
alpaca-trade-api>=2.0.0
pandas>=1.5.0
numpy>=1.21.0
pytz>=2022.7
aiohttp>=3.8.0
yfinance>=0.2.0
python-dateutil>=2.8.0
asyncio>=3.5.0